Move the projection fastpaths into projections.py
Pure move, no behavior change. projections.py holds raw-array projection code;
lib.py keeps the xarray-aware async orchestration around it.
