+++ src/xpublish_tiles/testing/datasets.py
@@ -14,8 +14,8 @@ from pyproj.aoi import BBox
 import dask.array
 import xarray as xr
 from xarray import DataTree
-from xpublish_tiles.lib import transformer_from_crs
 from xpublish_tiles.multiscale import assign_leaf_xpublish_ids
+from xpublish_tiles.projections import transformer_from_crs
 from xpublish_tiles.testing.tiles import (
     ETRS89_TILES,
     ETRS89_TILES_EDGE_CASES,
