+++ src/xpublish_tiles/xpublish/wms/utils.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import xarray as xr
 from xpublish_tiles.grids import guess_grid_system
-from xpublish_tiles.pipeline import transformer_from_crs
+from xpublish_tiles.projections import transformer_from_crs
 from xpublish_tiles.xpublish.wms.types import (
     WMSAttributeResponse,
     WMSBoundingBoxResponse,
